+++ tests/functional/test_install_report.py
@@ -12,7 +12,6 @@ def _install_dict(report: Dict[str, Any]) -> Dict[str, Any]:
     return {canonicalize_name(i["metadata"]["name"]): i for i in report["install"]}
 
 
-@pytest.mark.usefixtures("with_wheel")
 def test_install_report_basic(
     script: PipTestEnvironment, shared_data: TestData, tmp_path: Path
 ) -> None:
@@ -26,7 +25,6 @@ def test_install_report_basic(
         str(shared_data.root / "packages/"),
         "--report",
         str(report_path),
-        allow_stderr_warning=True,
     )
     report = json.loads(report_path.read_text())
     assert "install" in report
@@ -44,7 +42,6 @@ def test_install_report_basic(
     )
 
 
-@pytest.mark.usefixtures("with_wheel")
 def test_install_report_dep(
     script: PipTestEnvironment, shared_data: TestData, tmp_path: Path
 ) -> None:
@@ -59,7 +56,6 @@ def test_install_report_dep(
         str(shared_data.root / "packages/"),
         "--report",
         str(report_path),
-        allow_stderr_warning=True,
     )
     report = json.loads(report_path.read_text())
     assert len(report["install"]) == 2
@@ -68,7 +64,6 @@ def test_install_report_dep(
 
 
 @pytest.mark.network
-@pytest.mark.usefixtures("with_wheel")
 def test_install_report_index(script: PipTestEnvironment, tmp_path: Path) -> None:
     """Test report for sdist obtained from index."""
     report_path = tmp_path / "report.json"
@@ -78,7 +73,6 @@ def test_install_report_index(script: PipTestEnvironment, tmp_path: Path) -> Non
         "Paste[openid]==1.7.5.1",
         "--report",
         str(report_path),
-        allow_stderr_warning=True,
     )
     report = json.loads(report_path.read_text())
     assert len(report["install"]) == 2
@@ -99,7 +93,6 @@ def test_install_report_index(script: PipTestEnvironment, tmp_path: Path) -> Non
 
 
 @pytest.mark.network
-@pytest.mark.usefixtures("with_wheel")
 def test_install_report_vcs_and_wheel_cache(
     script: PipTestEnvironment, tmp_path: Path
 ) -> None:
@@ -114,7 +107,6 @@ def test_install_report_vcs_and_wheel_cache(
         str(cache_dir),
         "--report",
         str(report_path),
-        allow_stderr_warning=True,
     )
     report = json.loads(report_path.read_text())
     assert len(report["install"]) == 1
@@ -142,7 +134,6 @@ def test_install_report_vcs_and_wheel_cache(
         str(cache_dir),
         "--report",
         str(report_path),
-        allow_stderr_warning=True,
     )
     assert "Using cached pip_test_package" in result.stdout
     report = json.loads(report_path.read_text())
@@ -162,7 +153,6 @@ def test_install_report_vcs_and_wheel_cache(
 
 
 @pytest.mark.network
-@pytest.mark.usefixtures("with_wheel")
 def test_install_report_vcs_editable(
     script: PipTestEnvironment, tmp_path: Path
 ) -> None:
@@ -176,7 +166,6 @@ def test_install_report_vcs_editable(
         "#egg=pip-test-package",
         "--report",
         str(report_path),
-        allow_stderr_warning=True,
     )
     report = json.loads(report_path.read_text())
     assert len(report["install"]) == 1
@@ -189,7 +178,6 @@ def test_install_report_vcs_editable(
     assert pip_test_package_report["download_info"]["dir_info"]["editable"] is True
 
 
-@pytest.mark.usefixtures("with_wheel")
 def test_install_report_to_stdout(
     script: PipTestEnvironment, shared_data: TestData
 ) -> None:
@@ -203,11 +191,6 @@ def test_install_report_to_stdout(
         str(shared_data.root / "packages/"),
         "--report",
         "-",
-        allow_stderr_warning=True,
-    )
-    assert result.stderr == (
-        "WARNING: --report is currently an experimental option. "
-        "The output format may change in a future release without prior warning.\n"
     )
     report = json.loads(result.stdout)
     assert "install" in report