+++ tests/functional/test_install_reqs.py
@@ -62,7 +62,6 @@ def arg_recording_sdist_maker(
 
 
 @pytest.mark.network
-@pytest.mark.usefixtures("with_wheel")
 def test_requirements_file(script: PipTestEnvironment) -> None:
     """
     Test installing from a requirements file.
@@ -113,7 +112,6 @@ def test_schema_check_in_requirements_file(script: PipTestEnvironment) -> None:
         ("embedded_rel_path", True),
     ],
 )
-@pytest.mark.usefixtures("with_wheel")
 def test_relative_requirements_file(
     script: PipTestEnvironment, data: TestData, test_type: str, editable: bool
 ) -> None:
@@ -161,7 +159,6 @@ def test_relative_requirements_file(
 @pytest.mark.xfail
 @pytest.mark.network
 @need_svn
-@pytest.mark.usefixtures("with_wheel")
 def test_multiple_requirements_files(script: PipTestEnvironment, tmpdir: Path) -> None:
     """
     Test installing from multiple nested requirements files.
@@ -305,7 +302,7 @@ def test_install_local_with_subdirectory(script: PipTestEnvironment) -> None:
     result.assert_installed("version_subpkg.py", editable=False)
 
 
-@pytest.mark.usefixtures("enable_user_site", "with_wheel")
+@pytest.mark.usefixtures("enable_user_site")
 def test_wheel_user_with_prefix_in_pydistutils_cfg(
     script: PipTestEnvironment, data: TestData
 ) -> None:
@@ -330,33 +327,6 @@ def test_wheel_user_with_prefix_in_pydistutils_cfg(
     # Check that we are really installing a wheel
     assert "Running setup.py install for requiresupper" not in result.stdout
     assert "installed requiresupper" in result.stdout
-
-
-def test_install_option_in_requirements_file_overrides_cli(
-    script: PipTestEnvironment,
-    arg_recording_sdist_maker: Callable[[str], ArgRecordingSdist],
-) -> None:
-    simple_sdist = arg_recording_sdist_maker("simple")
-
-    reqs_file = script.scratch_path.joinpath("reqs.txt")
-    reqs_file.write_text("simple --install-option='-O0'")
-
-    result = script.pip(
-        "install",
-        "--no-index",
-        "-f",
-        str(simple_sdist.sdist_path.parent),
-        "-r",
-        str(reqs_file),
-        "--install-option=-O1",
-        allow_stderr_warning=True,
-    )
-    simple_args = simple_sdist.args()
-    assert "install" in simple_args
-    assert simple_args.index("-O1") < simple_args.index("-O0")
-    assert "Implying --no-binary=:all:" in result.stderr
-    assert "Consider using --config-settings" in result.stderr
-    assert "--install-option is deprecated" in result.stderr
 
 
 def test_constraints_not_installed_by_default(
@@ -493,7 +463,7 @@ def test_constraints_constrain_to_local(
         "singlemodule",
         allow_stderr_warning=True,
     )
-    assert "Running setup.py install for singlemodule" in result.stdout
+    assert "Building wheel for singlemodule" in result.stdout
 
 
 def test_constrained_to_url_install_same_url(
@@ -512,10 +482,9 @@ def test_constrained_to_url_install_same_url(
         to_install,
         allow_stderr_warning=True,
     )
-    assert "Running setup.py install for singlemodule" in result.stdout, str(result)
+    assert "Building wheel for singlemodule" in result.stdout, str(result)
 
 
-@pytest.mark.usefixtures("with_wheel")
 def test_double_install_spurious_hash_mismatch(
     script: PipTestEnvironment, tmpdir: Path, data: TestData
 ) -> None:
@@ -644,7 +613,7 @@ def test_install_distribution_full_union(
     result = script.pip_install_local(
         to_install, f"{to_install}[bar]", f"{to_install}[baz]"
     )
-    assert "Running setup.py install for LocalExtras" in result.stdout
+    assert "Building wheel for LocalExtras" in result.stdout
     result.did_create(script.site_packages / "simple")
     result.did_create(script.site_packages / "singlemodule.py")
 
@@ -759,61 +728,3 @@ def test_install_unsupported_wheel_file(
         in result.stderr
     )
     assert len(result.files_created) == 0
-
-
-def test_install_options_local_to_package(
-    script: PipTestEnvironment,
-    arg_recording_sdist_maker: Callable[[str], ArgRecordingSdist],
-) -> None:
-    """Make sure --install-options does not leak across packages.
-
-    A requirements.txt file can have per-package --install-options; these
-    should be isolated to just the package instead of leaking to subsequent
-    packages.  This needs to be a functional test because the bug was around
-    cross-contamination at install time.
-    """
-
-    simple1_sdist = arg_recording_sdist_maker("simple1")
-    simple2_sdist = arg_recording_sdist_maker("simple2")
-
-    reqs_file = script.scratch_path.joinpath("reqs.txt")
-    reqs_file.write_text(
-        textwrap.dedent(
-            """
-            simple1 --install-option='-O0'
-            simple2
-            """
-        )
-    )
-    script.pip(
-        "install",
-        "--no-index",
-        "-f",
-        str(simple1_sdist.sdist_path.parent),
-        "-r",
-        reqs_file,
-        allow_stderr_warning=True,
-    )
-
-    simple1_args = simple1_sdist.args()
-    assert "install" in simple1_args
-    assert "-O0" in simple1_args
-    simple2_args = simple2_sdist.args()
-    assert "install" in simple2_args
-    assert "-O0" not in simple2_args
-
-
-def test_location_related_install_option_fails(script: PipTestEnvironment) -> None:
-    simple_sdist = create_basic_sdist_for_package(script, "simple", "0.1.0")
-    reqs_file = script.scratch_path.joinpath("reqs.txt")
-    reqs_file.write_text("simple --install-option='--home=/tmp'")
-    result = script.pip(
-        "install",
-        "--no-index",
-        "-f",
-        str(simple_sdist.parent),
-        "-r",
-        reqs_file,
-        expect_error=True,
-    )
-    assert "['--home'] from simple" in result.stderr